* Issue 1: Episode ordering other than "aired" broken
Changed episode thumbnail name when downloading different orders other than 'Aired', this will fix issue when wrong thumbnail is displayed when user removes database entry for aired series and re-imports as different order e.g. DVD

Note: When deleting the series and re-importing as a different order the thumbnail wont download again unless you exit the program and re-enter then run import...reason being is that the series is cached with the older order. This could be fixed but its a minor issue since the user needs to exit and re-enter the plugin anyway