* Issue 70: Fanart Chooser selects first thumbnail item upon download complete.
It now remembers the last selected item before it re-loads the facade...can be a bit jerky when moving around and multiple downloads finishing/reloading. 

I think to speed it up, reduce the size of the images being added to the thumbnails (as these are generally a fraction of the fanart size) and/or load them in another thread like in the TV series GUI (but that's a bit overkill for the fanart chooser).