Fixed local_series name being used for thumbs directory instead of the Pretty Name.
Note: The way the cache works still needs to be looked at / repaired.  It is caching things too early ,it seems.  So on a new series, it puts everything in the local_series name because it doesn't know the Pretty Name yet.  i.e. cached before updated online?


Grammar fix in GUIConfiguration.designer.cs