Tweaked Parental Controls.
There appears to be a an issue with MediaPortal when 'Show Last Active Module' is enabled. If this option is enabled and the last active module was TVSeries on a protected view, when you restart MediaPortal this will invoke the PinCode dialog...

Problem is MediaPortal will not completely start, it doesn't show the Modal Dialog. Im pretty sure this is a MediaPortal issue as it happens with all plugins (simply delete the skin file on your last active plugin and restart..the skin warning dialog will not appear (tested with movpics and myvideos)!)

Either need to find a work around in this scenario or fallback to an unprotected view or ...