Updated Facade colors to use <remoteColor> and <playedColor>, this uses the corresponding GUIListItem properties IsRemote and IsPlayed.
Note: IsRemote takes precedence so is evaluated first then IsPlayed. If both are false then the color is set to <textColor>.

Commented out the color overrides in the Blue3 skins so that it now uses the references.xml. Users can edit the file if they want different colors. I think its makes a lot more sense now when editing the colors, I was always confused on what color is what unless I read the code...it should be pretty obvious now.