Issue 339: Fixed Series/Season Remote/Offline Color and Episode Counts do not update when deleting from disk from GUI.
This would be a good idea to extract the 'Cleanup/Facade Handling' out of the delete methods into a new method which is called after all the delete operations are done.

Also would be nice to have a GUIProgress dialog, would be esp useful for deletes at the series/season level on a NAS which can sometimes take a long time.