Issue 450: Fixed issues with local file references in episode table being removed when import path is not available. Removed 'keep_references' and 'removable' properties, these are handled internally.
'Keep references' will be done if a file does not exist and the corresponding import path does not exist...all other cases the file will be removed as we a certain its been physically removed.

'Removable' is not needed, if the import path for the file does not exist whether it be a USB Disk, CD/DVD or network drive, it will be prompted to insert media when attempted to play.

Issue 466: Fixed OnVolumeInserted/Removed event when import path was entered as lower-case.

Issue 468: Fixed duplicate episodes added when using Mapped Drive letter for ImportPath and File Monitor picks up the same file i.e. previous time network was not available.