Added support to retrieve community ratings from trakt.tv for shows and episodes. There is still some tweaking to be done when to retrieve ratings from online, its very conservative right now. Some changes are needed for when new episodes are added to database to ensure ratings are added in that case.
A new field has been added to the series table to store the Trakt ID, the first time the import is run it will be slower as it needs to retrieve these IDs. 

Note: this requires the next version of the trakt plugin. If testing, you can use the dll in the Externals folder.